requirements.txt
python-dotenv
requests

google-api-python-client==2.143.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
python-dateutil==2.9.0.post0

asyncio

openai-agents

streamlit